+++ example/domain/uno/mongodb/spring/blocking/src/main/kotlin/org/occurrent/example/domain/uno/es/spring/blocking/Bootstrap.kt
@@ -19,7 +19,7 @@ import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import org.occurrent.application.converter.CloudEventConverter
 import org.occurrent.application.service.blocking.execute
-import org.occurrent.application.service.blocking.implementation.GenericApplicationService
+import org.occurrent.application.service.blocking.generic.GenericApplicationService
 import org.occurrent.eventstore.api.WriteConditionNotFulfilledException
 import org.occurrent.eventstore.api.blocking.EventStoreQueries
 import org.occurrent.eventstore.mongodb.spring.blocking.EventStoreConfig